[ALPS03938839] build: fix implicit conversion from int to short
Fix the clang build error:
error: implicit conversion from int to short changes
value from 32769 to -32767 [-Werror,-Wconstant-conversion]

MTK-Commit-Id: 4d34b04f1a00d689d0b73833ab01706e26c0a445

CR-Id: ALPS03938839
Feature: [Module]Kernel Maintenance

+++ drivers/misc/mediatek/m4u/2.4/m4u_mva.c
@@ -34,7 +34,7 @@ void m4u_mvaGraph_init(void *priv_reserve)
 	spin_lock_irqsave(&gMvaGraph_lock, irq_flags);
 	memset(mvaGraph, 0, sizeof(short) * (MVA_MAX_BLOCK_NR + 1));
 	memset(mvaInfoGraph, 0, sizeof(void *) * (MVA_MAX_BLOCK_NR + 1));
-	mvaGraph[0] = 1 | MVA_BUSY_MASK;
+	mvaGraph[0] = 1 | (short)MVA_BUSY_MASK;
 	mvaInfoGraph[0] = priv_reserve;
 	/*need to reserve two mva block region for vpu.*/
 	/* *[1, 0x500): shared with all ports.*/
